pyproject.toml
[tool.poetry]
name = "kaggle_panda"
version = "0.1.0"
description = ""
authors = ["MasatakaImamura <masataka.i.navi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
numpy = "^1.18.5"
pandas = "^1.0.5"
matplotlib = "^3.2.2"
seaborn = "^0.10.1"
scipy = "^1.4.1"
scikit-learn = "^0.23.1"
tqdm = "^4.46.1"
opencv-python = "^4.2.0"
tensorboardx = "^2.0"
tensorboard = "^2.2.2"
pillow = "^7.1.2"
albumentations = "^0.4.5"
mlflow = "^1.9.1"
hydra-core = "^0.11.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
